ADSSDK-10

Approved-by: Blagojco Juca

+++ insidead/src/main/java/com/streann/insidead/callbacks/InsideAdCallback.kt
@@ -14,6 +14,6 @@ interface InsideAdCallback {
 
     fun insideAdError(error: String)
 
-    fun insideAdVolumeChanged(level: Float)
+    fun insideAdVolumeChanged(level: Int)
 
 }
